+++ src/super_gradients/training/datasets/detection_datasets/detection_dataset.py
@@ -73,7 +73,7 @@ class DetectionDataset(Dataset):
         cache: bool = False,
         cache_dir: str = None,
         transforms: List[DetectionTransform] = [],
-        all_classes_list: Optional[List[str]] = None,
+        all_classes_list: Optional[List[str]] = [],
         class_inclusion_list: Optional[List[str]] = None,
         ignore_empty_annotations: bool = True,
         target_fields: List[str] = None,
@@ -109,6 +109,7 @@ class DetectionDataset(Dataset):
 
         # Number of images that are available (regardless of ignored images)
         self.n_available_samples = self._setup_data_source()
+
         if not isinstance(self.n_available_samples, int) or self.n_available_samples < 1:
             raise ValueError(f"_setup_data_source() should return the number of available samples but got {self.n_available_samples}")
 
@@ -122,10 +123,10 @@ class DetectionDataset(Dataset):
         if class_inclusion_list is not None and len(class_inclusion_list) != len(set(class_inclusion_list)):
             raise DatasetValidationException(f"class_inclusion_list contains duplicate class names: {collections.Counter(class_inclusion_list)}")
 
-        self.all_classes_list = all_classes_list
+        self.all_classes_list = all_classes_list or self._all_classes
         self.class_inclusion_list = class_inclusion_list
         self.classes = self.class_inclusion_list or self.all_classes_list
-        if len(set(self.classes) - set(all_classes_list)) > 0:
+        if len(set(self.classes) - set(self.all_classes_list)) > 0:
             wrong_classes = set(self.classes) - set(all_classes_list)
             raise DatasetValidationException(f"class_inclusion_list includes classes that are not in all_classes_list: {wrong_classes}")
 
@@ -146,6 +147,11 @@ class DetectionDataset(Dataset):
         self.output_fields = output_fields or ["image", "target"]
         if len(self.output_fields) < 2 or self.output_fields[0] != "image" or self.output_fields[1] != "target":
             raise ValueError('output_fields must start with "image" and then "target", followed by any other field')
+
+    @property
+    def _all_classes(self):
+        """Placeholder to dynamically setup the class names. This is prefered over passing all_classes_list to __init__."""
+        raise NotImplementedError
 
     def _setup_data_source(self) -> int:
         """Set up the data source and store relevant objects as attributes.
